utils: factorize and remove global variables
* squash vdev_constants.py and utils.py in vdev_utils.py. These two
files are quiet small: three constants in the first one and two methods
in the second, So, this makes no sense to keep these files as separate
sources, at least at the current moment.

* vdev_utils.get_vdev_cmd() method is called in networking-6wind and
os-vif-vhost-user modules. So, there is no more need to set global
variables: FP_VDEV_CMD and FP_VDEV_CMD_CHECK in fp_vdev_remote
namespace. These global variables should be now used in networking-6wind
and os-vif-vhost-user namespaces.

* vdev_utils.command_exists(): just return True or False, if a given
binary (fp-vdev/fp-vdev-remote) is existed or absent. Don't return an
absolute path to it, this allows to avoid random encode/decode errors in
oslo_rootwrap.execute(), when we pass a string with the absolute path to
the fp-vdev/fp-vdev-remote as an argument.

Signed-off-by: Valentina Krasnobaeva <valentina.krasnobaeva@6wind.com>
Acked-by: Maxime Leroy <maxime.leroy@6wind.com>